lora_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_sysroot_linux-64_curr_repodata_hack=3=haa98f57_10
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.4.0=pypi_0
accelerate=0.21.0=pyhd8ed1ab_0
aiohttp=3.8.5=py39h5eee18b_0
aiosignal=1.3.1=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
arrow=1.2.3=py39h06a4308_1
arrow-cpp=11.0.0=h374c478_2
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
audioread=3.0.0=py39hf3d152e_1
aws-c-common=0.6.8=h5eee18b_1
aws-c-event-stream=0.1.6=h6a678d5_6
aws-checksums=0.1.11=h5eee18b_2
aws-sdk-cpp=1.8.185=h721c034_1
binaryornot=0.4.4=pyhd3eb1b0_1
binutils_impl_linux-64=2.38=h2a08ee3_1
binutils_linux-64=2.38.0=hc2dff05_0
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_1
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.11.17=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.15.1=py39h5eee18b_3
chardet=4.0.0=py39h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=py39hdb19cb5_0
cookiecutter=1.7.3=pyhd3eb1b0_0
cryptography=41.0.2=py39h22a60cf_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvcc=11.7.99=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd3eb1b0_0
cython=3.0.2=pypi_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.15.0=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
diffusers=0.15.0=pypi_0
dill=0.3.7=pypi_0
evaluate=0.4.0=py39h06a4308_0
fairseq=0.12.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fsspec=2023.9.2=pypi_0
gcc_impl_linux-64=11.2.0=h1234567_1
gcc_linux-64=11.2.0=h5c386dc_0
gettext=0.21.1=h27087fc_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpc-cpp=1.48.2=he1ff14a_1
grpcio=1.57.0=pypi_0
gxx_impl_linux-64=11.2.0=h1234567_1
gxx_linux-64=11.2.0=hc2dff05_0
huggingface_hub=0.15.1=py39h06a4308_0
icu=73.1=h6a678d5_0
idna=3.4=py39h06a4308_0
importlib-metadata=6.8.0=pypi_0
importlib_metadata=7.0.0=hd8ed1ab_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py39h06a4308_0
jinja2-time=0.2.0=pyhd3eb1b0_3
jiwer=3.0.3=pypi_0
joblib=1.2.0=py39h06a4308_0
jpeg=9e=h5eee18b_1
kernel-headers_linux-64=3.10.0=h57e8cba_10
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy_loader=0.2=py39h06a4308_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=1_h86c2bf4_netlib
libboost=1.82.0=ha8e66a6_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcblas=3.9.0=5_h92ddd45_netlib
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcurl=8.2.1=h251f7ec_0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libflac=1.4.3=h59595ed_0
libgcc-devel_linux-64=11.2.0=h1234567_1
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libgomp=13.1.0=he5830b7_0
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
liblapack=3.9.0=5_h92ddd45_netlib
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.52.0=h2d74bed_1
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
librosa=0.10.1=pyhd8ed1ab_0
libsndfile=1.2.2=hbc2eb40_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-devel_linux-64=11.2.0=h1234567_1
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
llvmlite=0.40.0=py39he621ea3_0
lxml=4.9.3=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-base=3.7.2=py39h1128e8f_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpg123=1.31.3=hcb278e6_0
mpmath=1.3.0=py39h06a4308_0
msgpack-python=1.0.3=py39hd09550d_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
npy-append-array=0.9.16=pyhd8ed1ab_0
numba=0.57.1=py39hb75a051_0
numexpr=2.8.4=py39hc78ab66_1
numpy=1.22.3=py39hc58783e_2
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.2.0=hd590300_1
orc=1.7.4=hb3bc3d3_1
packaging=23.1=py39h06a4308_0
pandas=2.0.3=py39h1128e8f_0
peft=0.5.0=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.2.1=py39h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
poyo=0.5.0=pyhd3eb1b0_0
protobuf=3.20.1=pypi_0
psutil=5.9.0=py39h5eee18b_0
pyarrow=11.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
pysoundfile=0.12.1=pyhd8ed1ab_0
python=3.9.17=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py39h5eee18b_1
python_abi=3.9=2_cp39
pytorch=2.0.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2022.7=py39h06a4308_0
pyyaml=6.0=py39h5eee18b_1
rapidfuzz=3.5.2=pypi_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2023.8.8=pypi_0
requests=2.31.0=py39h06a4308_0
requests-oauthlib=1.3.1=pypi_0
responses=0.13.3=pyhd3eb1b0_0
rsa=4.9=pypi_0
sacremoses=0.0.53=pyhd8ed1ab_0
safetensors=0.3.3=py39h9fdd4d6_1
scikit-learn=1.3.0=py39h1128e8f_0
scipy=1.11.1=py39h5f9d8c6_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
soxr=0.1.3=h0b41bf4_3
soxr-python=0.3.5=py39h0f8d45d_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py39h06a4308_0
sysroot_linux-64=2.17=h57e8cba_10
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
text-unidecode=1.3=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=py39h22610ee_0
torchaudio=2.0.2=py39_cu117
torchtriton=2.0.0=py39
torchvision=0.15.2=py39_cu117
tqdm=4.66.1=pypi_0
transformers=4.30.1=pypi_0
typing-extensions=4.7.1=py39h06a4308_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.16=py39h06a4308_0
utf8proc=2.6.1=h27cfd23_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py39h06a4308_0
xxhash=3.3.0=pypi_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.2=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
